B - sfhash_hashset_builder_write() should return a uint64_t, not a size_t, to prevent truncation on 32-bit systems.

+++ include/hasher/hashset.h
@@ -125,7 +125,7 @@ void sfhash_hashset_builder_add_hash(
   size_t length
 );
 
-size_t sfhash_hashset_builder_write(
+uint64_t sfhash_hashset_builder_write(
   SFHASH_HashsetBuildCtx* bctx,
   SFHASH_Error** err
 );
